requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
greenlet==2.0.2
gunicorn==23.0.0
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
SQLAlchemy==2.0.47
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
URLObject==3.0.0
Werkzeug==3.1.6
zipp==3.23.0
